+++ src/index.html
@@ -16,7 +16,6 @@
   <body >
     <!-- Start Main -->
     <div id="body">
-
     </div>
     <main >
       <ul class="f-list | row list-unstyled mt-5 p-5 mx-auto"></ul>
