+++ src/modules/comments.js
@@ -1,0 +1,22 @@
+
+
+const comments = ( comm, data) => {
+  console.log(data, comm)
+  data.map((datd) => {
+    const dateComment1 = document.createElement('div');
+    dateComment1.setAttribute('class', 'date_comment');
+    const b1 = document.createElement('b');
+    b1.setAttribute('class', 'detail_p');
+    b1.innerText = `${datd.creation_date}`;
+    const p5 = document.createElement('p');
+    p5.setAttribute('class', 'detail_p');
+    p5.innerText = `${datd.username}: ${datd.comment}`;
+    dateComment1.appendChild(b1);
+    dateComment1.appendChild(p5);
+    comm.appendChild(dateComment1);
+
+    return null;
+  });
+}
+
+export default comments;
